# RFC: handle serializing objects with cycles using a Serializer object