--- a/base/exports.jl
+++ b/base/exports.jl
@@ -97,6 +97,7 @@ export
     RoundNearestTiesUp,
     RoundToZero,
     RoundUp,
+    SerializationState,
     Set,
     SharedArray,
     SharedMatrix,