Been meaning to write this up for a while.

Serializing now uses a `Serializer` object that keeps the needed state to handle cycles. In some cases it also handles repeated references to the same object.

As in my last attempt, when defining `serialize` you use

```
    serialize_cycle(s, obj) && return
```

and when defining `deserialize` you use

```
    deserialize_cycle(s, newobj, pos)
```