@JeffBezanson 2015-02-12
@amitmurthy What's the best way to combine this with BufferedAsyncStream? We need `position` in order to do back-references.