@amitmurthy 2015-02-13
As a standalone implementation,  `BufferedAsyncStream` can be modified to use a seekable buffer. It seemed more natural to use a `PipeBuffer` since it was an AsyncStream.

Same thing if we decide to move the implementation of `BufferedAsyncStream` into all `AsyncStream` types. AsyncStream's are not seekable.