@JeffBezanson 2015-02-13
I don't need seeking for this, just `position`. We could keep a running byte counter in [Buffered]AsyncStreams and use offsets relative to that for each message.