@amitmurthy 2015-02-13
The recv buffer in `TCPSocket` as well the send buffer in `BufferedAsyncStream` are bypassed for large arrays. The send buffer is also flushed if its size becomes greater that 100K. Wouldn't that be a problem for a Serializer object? They are also not aware of any "message" construct - just a stream of bytes.