@JeffBezanson 2015-02-13
No, as I said the Serializer does not need to seek or re-read the buffer. It keeps a table mapping offsets to objects, so it can look up already-(de)serialized objects when it sees a back-reference later in the stream.

However some kind of message boundary is needed, because there has to be some limit to what a back-reference can refer to. The end of a "message" (however defined) is when it is OK to clear the back-reference table. Messages can be implemented on top of the stream, they don't have to be part of the stream object. Basically they need to be managed by whoever calls `serialize` and `deserialize`.