@amitmurthy 2015-02-15
- Since serialize/deserialize needs to work across all types of IO, and async streams do not have a concept of a `position` or `reset`, we could add counters to `Serializer` itself and track bytes recd/sent via read/write calls.

- `reset` on a `Serializer` object will clear all state. We should also have a member auto-reset value. This would avoid inadvertent memory leaks. via user code that inadvertantly does not call `reset`.

- As issue with using a Serializer as the only interface to serialize/deserialize would be that mixing in-memory buffers and tcpsockets/file io could become complicated. For example, currently, it is straightforward to

    - serialize some objects to an IOBuffer (to avoid a large number of syscalls)
    - takebuffer, write to fd, serialize a large object directly to fd
    - serialize some more to IOBuffer
    - finally takebuffer, write to fd and flush

- We could probably layer the `serialize(s::Serializer, ....)` methods above regular `serialize(s::IO, ....)` methods. Only user code that needs the specific optimization provided by a Serializer object can choose to use it via a copy constructor (to set different fds) to track the same counters where applicable.

- There could be other scenarios where serializing is done across both in-memory buffers and system IO calls. I cannot think of a good situation and I am not sure if it is a good idea to have two interfaces (for e.g., bugs dues to serializing via a Serializer and restore directly from fd), but we probably should.

- All IO, wherever applicable should support a send buffer. I'll start by submitting a PR for (as discussed towards the end of #10073)  merging the functionality of `BufferedAsyncStream` into all `AsyncStreams`.

- `Serializer` could also implement the functionality of `BufferedAsyncStream` (for serialize only). AsyncStreams already have a recv buffer. Not really in favor of this though.