@JeffBezanson 2015-02-20
I have changed it not to require `position`. The parallel test now passes for me, so we could use this implementation and leave further issues for later. Currently you can still simply call `serializer(io, x)` and it will work no worse than before, except also handle cycles within `x`.