@vtjnash 2015-02-20
does not specializing `Serializer` on the type of `io` have any performance implications relative to the previous incarnation?

should `serialize_cycle` handle the `isimmutable` check instead of the callsites? also, it perhaps shouldn't branch on `isimmutable` as you do now, but on `isbits` (for example, an immutable containing an array containing itself seems like it might be suboptimal, no?)

the `lambda_numbers` and `known_lambda_data ` dictionaries should go away