@JeffBezanson 2015-02-20
We should look at performance. It would be nice if we didn't have to specialize I/O code on every kind of stream.

You're right about `isbits`.

What would be another way to avoid recompiling when the same function is sent repeatedly?