@vtjnash 2015-02-20
Either Serializer state or serialize_cycle