@vtjnash 2015-02-20
More comments: the keys of the table would ideally be WeakRef objects

The lazy initialization of the table feels like a premature (de)optimization, since it should almost always be getting created