@StefanKarpinski 2015-02-20
I like this approach. I think this transparently handles 99% of what people want. If you want to share state across the serialization of multiple objects do you just manually construct a `Serializer` object and then serialize to that instead of to a bare `IO` object? There's a bit of an issue with that in that the receiving side then would need to have a Serializer object with the exact same lifetime, which seems a little tricky to orchestrate. But maybe that's an advanced use-case and we can assume anyone doing that knows what they're doing.