@vtjnash 2015-02-20
I thought "having the same serializer state" was an implied given. We could now even have Serializer write a header to verify this. The lifetime bound tends to be pretty simple (likely just the same as the underlying stream)