@amitmurthy 2015-02-21
The lifetime bound is important when you think of long held streams - think `multi.jl`. This is when the same serializer state on either side becomes important and we will need some sort of a periodic reset (to avoid unnecessarily holding references).

A `reset` tag sent on the stream (whenever either side calls a `reset` method on the Serializer), should handle it.

As suggested by Jeff, currently it handles the case of self-referencing objects, the optimisation part for long held streams and sending the same object repeatedly (for e.g., darrays being sent across mutiple times) can be discussed separately.