@vtjnash 2015-02-21
a `reset` tag sounds very finicky. i think it would do much better to periodically sweep all of the `WeakRef` objects in the `table` (perhaps a callback finalizer from the gc), and serializing to the stream a list of any that can be discarded on the other end.

but i guess this can merge and deal with the other optimizations later. as it stands, this is very simply a drop-in replacement of the existing deserializer and can be merged as such