@JeffBezanson 2015-02-25
Ok I checked #7893 with this, and the performance is simply awful. Will investigate.