@JeffBezanson 2015-03-10
Ok the performance is no longer awful. Just a few percent slower than what we have now.