@JeffBezanson 2015-03-18
Status of performance with a 240MB DataFrame:

On master:

```
julia> @time d = deserialize(open("/home/jeff/src/julia/ioe.jld"))
elapsed time: 13.120145274 seconds (1158 MB allocated, 2.07% gc time in 6 pauses with 4 full sweep)

julia> @time serialize(open("out.jld","w"), d);
elapsed time: 7.78289567 seconds (643 MB allocated, 5.31% gc time in 1 pauses with 1 full sweep)
```

on this branch:

```
julia> @time d = deserialize(open("/home/jeff/src/julia/ioe.jld"))
elapsed time: 14.836951264 seconds (1162 MB allocated, 2.04% gc time in 7 pauses with 4 full sweep)

julia> @time serialize(open("out.jld","w"), d);
elapsed time: 9.000394767 seconds (650 MB allocated, 4.63% gc time in 1 pauses with 1 full sweep)
```