@JeffBezanson 2015-03-18
Ok, back to parity with master for lots of pointer-free objects:

```
julia> @time d = deserialize(open("/home/jeff/src/julia/ioe.jld"));
elapsed time: 12.809747142 seconds (1161 MB allocated, 2.35% gc time in 7 pauses with 4 full sweep)

julia> @time serialize(open("out.jld","w"), d);
elapsed time: 7.926651548 seconds (650 MB allocated, 5.35% gc time in 1 pauses with 1 full sweep)
```

But, for large graphs of mutable objects the performance can be really bad. We might need an option to disable cycle support.